+++ yux/ast/ast_builder.h
@@ -108,7 +108,6 @@ public:
     std::any visitExprBool(yux::yuxParser::ExprBoolContext* ctx) override;
     std::any visitExprIfElse(yux::yuxParser::ExprIfElseContext* ctx) override;
     std::any visitExprOneLineIfElse(yux::yuxParser::ExprOneLineIfElseContext* ctx) override;
-    std::any visitExprIfElsePreValue(yux::yuxParser::ExprIfElsePreValueContext* ctx) override;
     std::any visitExprElIf(yux::yuxParser::ExprElIfContext* ctx) override;
     std::any visitExprElse(yux::yuxParser::ExprElseContext* ctx) override;
     std::any visitExprGet(yux::yuxParser::ExprGetContext* ctx) override;